# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Synvet — Plataforma SaaS Veterinária
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ConsultationWithPet" }
    post:
      operationId: createConsultation
      tags: [consultations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateConsultationBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Consultation" }
  /consultations/{consultationId}:
    parameters:
      - in: path
        name: consultationId
        required: true
        schema: { type: string }
    get:
      operationId: getConsultation
      tags: [consultations]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsultationDetail" }
    patch:
      operationId: updateConsultation
      tags: [consultations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateConsultationBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Consultation" }
    delete:
      operationId: deleteConsultation
      tags: [consultations]
      responses:
        "204":
          description: No Content
  /consultations/{consultationId}/anamnesis:
    parameters:
      - in: path
        name: consultationId
        required: true
        schema: { type: string }
    get:
      operationId: getAnamnesis
      tags: [anamneses]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Anamnesis" }
    put:
      operationId: upsertAnamnesis
      tags: [anamneses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertAnamnesisBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Anamnesis" }
...
components:
  schemas:
...
    Tutor:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        whatsapp: { type: string, nullable: true }
        address: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, createdAt, updatedAt]
...
    Pet:
      type: object
      properties:
        id: { type: string }
        tutorId: { type: string }
        name: { type: string }
        species: { type: string }
        breed: { type: string, nullable: true }
        sex: { type: string, enum: [male, female, unknown] }
        birthDate: { type: string, format: date, nullable: true }
        weightKg: { type: number, nullable: true }
        neutered: { type: boolean }
        allergies: { type: string, nullable: true }
        continuousMedications: { type: string, nullable: true }
        isCritical: { type: boolean }
        notes: { type: string, nullable: true }
        photoUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, tutorId, name, species, sex, neutered, isCritical, createdAt, updatedAt]
...
    Consultation:
      type: object
      properties:
        id: { type: string }
        petId: { type: string }
        scheduledAt: { type: string, format: date-time }
        status: { type: string, enum: [scheduled, in_progress, completed, cancelled] }
        reason: { type: string, nullable: true }
        symptoms: { type: string, nullable: true }
        observations: { type: string, nullable: true }
        evolution: { type: string, nullable: true }
        medications: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, petId, scheduledAt, status, createdAt, updatedAt]
    ConsultationWithPet:
      allOf:
        - $ref: "#/components/schemas/Consultation"
        - type: object
          properties:
            petName: { type: string }
            petSpecies: { type: string }
            tutorName: { type: string }
          required: [petName, petSpecies, tutorName]
    ConsultationDetail:
      allOf:
        - $ref: "#/components/schemas/Consultation"
        - type: object
          properties:
            pet: { $ref: "#/components/schemas/Pet" }
            tutor: { $ref: "#/components/schemas/Tutor" }
            anamnesis: { $ref: "#/components/schemas/Anamnesis", nullable: true }
          required: [pet, tutor]
    CreateConsultationBody:
      type: object
      properties:
        petId: { type: string }
        scheduledAt: { type: string, format: date-time }
        status: { type: string, enum: [scheduled, in_progress, completed, cancelled] }
        reason: { type: string, nullable: true }
        symptoms: { type: string, nullable: true }
        observations: { type: string, nullable: true }
        evolution: { type: string, nullable: true }
        medications: { type: string, nullable: true }
      required: [petId, scheduledAt, status]
    UpdateConsultationBody:
      type: object
      properties:
        scheduledAt: { type: string, format: date-time }
        status: { type: string, enum: [scheduled, in_progress, completed, cancelled] }
        reason: { type: string, nullable: true }
        symptoms: { type: string, nullable: true }
        observations: { type: string, nullable: true }
        evolution: { type: string, nullable: true }
        medications: { type: string, nullable: true }
    Anamnesis:
      type: object
      properties:
        id: { type: string }
        consultationId: { type: string }
        neurological: { type: string, nullable: true }
        digestive: { type: string, nullable: true }
        respiratory: { type: string, nullable: true }
        dermatological: { type: string, nullable: true }
        general: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, consultationId, createdAt, updatedAt]
    UpsertAnamnesisBody:
      type: object
      properties:
        neurological: { type: string, nullable: true }
        digestive: { type: string, nullable: true }
        respiratory: { type: string, nullable: true }
        dermatological: { type: string, nullable: true }
        general: { type: string, nullable: true }
...